[REFEAT]: LeftView, RightView 설정 관련 코드 Core로 이동
